refactor(frontend/src/app/layout/Navbar.jsx):
move NavBar component to app/layout/Navbar.jsx for layout organization
