+++ frontend/src/app/App.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { getCommand, getCommands } from "../features/commands/api/apiCommands";
-import { Navbar } from "../components/navbar/Navbar";
+import { Navbar } from "./layout/Navbar";
 import { Route, Routes } from "react-router-dom";
 import { Home } from "../features/commands/lookup/Home";
 import { FilterCmd } from "../components/pages/Filter/FilterCmd";
